Open overlay menu with ON key instead of back button
Hooking the back button is considered bad behavior and unintuitive.
Use the ON key for opening the ROM overlay menu. It is the only key that
is not used frequently and can be replaced by the an entry in the menu.

Also, as this is an Android application, we don't need to "turn off" the
calculator to save battery; we just leave the activity instead.

While we're at it, also remove the (kind of) unused volume button hooking.
We may bring that back later via a proper configuration option.