Update target sdk and build dependencies
* Target SDK 29
* Gradle 7.0.2
* Gson 2.7

We can't update the android support library v4 past 24.1.1 because this
was the last version that supports a minSdkVersion of 8.
Also, we need to target SDK version 29 instead of 30 for now as we need
requestLegacyExternalStorage which is not available in 30+ anymore.
This may be circumvented if we move to the newer storage access methods.